requirements.txt
chaind<=0.0.1,>=0.0.1a3
hexathon~=0.0.1a7
chainlib-eth<=0.0.5,>0.0.4
